requirements.txt
aiogram
google-generativeai
python-dotenv